agile-sprint-assistant/requirements.txt
google-generativeai==0.3.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2